Create Relu_Funct.py
In this example, we define a function relu() that takes an input x and applies the ReLU activation function to it. The ReLU function simply returns the input value if it's positive, and 0 otherwise.

We use the NumPy library to apply the ReLU function to an array of input values x. The np.maximum() function takes two arrays as input and returns an array where each element is the maximum of the corresponding elements in the input arrays. In this case, we use np.maximum() to return an array where each element is either the input value x or 0 (if the input value is negative).

Finally, we print the output array y to see the result of applying the ReLU function to the input array x.